+++ src/FileBgen.cpp
@@ -7,8 +7,8 @@ void FileBgen::read_all_and_centering()
     cout << timestamp() << "begin to parse the bgen file.\n";
     if (!params.pcangsd)
     {
-        F = VectorXf::Zero(nsnps);
-        G = MatrixXf(nsamples, nsnps);
+        F = VectorXd::Zero(nsnps);
+        G = MatrixXd(nsamples, nsnps);
         if (params.runem) C.resize(nsnps * nsamples);
         for (j = 0; j < nsnps; j++) {
             try {
@@ -60,8 +60,8 @@ void FileBgen::read_all_and_centering()
         }
         assert( j == nsnps );
         cout << timestamp() << "begin to estimate allele frequencies using GP" << endl;
-        VectorXf Ft(nsnps);
-        F = VectorXf::Constant(nsnps, 0.25);
+        VectorXd Ft(nsnps);
+        F = VectorXd::Constant(nsnps, 0.25);
         // run EM to estimate allele frequencies
         double diff;
         for (uint it = 0; it < params.maxiter; it++)
@@ -89,7 +89,7 @@ void FileBgen::read_all_and_centering()
             }
         }
         // initial E which is G
-        G = MatrixXf(nsamples, nsnps);
+        G = MatrixXd(nsamples, nsnps);
         #pragma omp parallel for
         for (j = 0; j < nsnps; j++) {
             double p0, p1, p2;
@@ -109,7 +109,7 @@ void FileBgen::read_snp_block_initial(uint64 start_idx, uint64 stop_idx, bool st
     uint i, j, snp_idx;
     if (G.cols() < params.blocksize || (actual_block_size < params.blocksize))
     {
-        G = MatrixXf::Zero(nsamples, actual_block_size);
+        G = MatrixXd::Zero(nsamples, actual_block_size);
     }
     if (frequency_was_estimated)
     {
